move pytest fixtures and config to a plugin
