+++ docs/testing.rst
@@ -73,10 +73,9 @@ The entrypoint waiter is a context manager that does not exit until a named entr
 Note that the context manager waits not only for the entrypoint method to complete, but also for any dependency teardown. Dependency-based loggers such as (TODO: link to bundled logger) for example would have also completed.
 
 
-Minimal conftest.py
--------------------
+Using py.test
+-------------
 
-Below is a minimal conftest for pytest that includes configuration and fixtures for communicating with RabbitMQ as well as the eventlet monkey patch. It is a good base for service tests if you're using pytest.
+Nameko's test suite uses py.test, and makes some useful configuration and fixtures available for your own tests if you choose to use pytest.
 
-.. literalinclude:: examples/testing/conftest.py
-
+They are contained in :mod:`nameko.testing.pytest`. This module is `automatically registered as a pytest plugin <https://pytest.org/latest/plugins.html#setuptools-entry-points>`_ by setuptools if you have pytest installed.
